# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Panel API",
        "description": "This application is developed by full stack web developer Tugrul Yildirim",
        "contact": {
            "email": "contact@tugrulyildirim.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://admin.local/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "register"
                ],
                "summary": "Register a new user",
                "description": "Returns the user data",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserAuthRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAuthResponseSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Login a user",
                "description": "Returns the user data",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserAuthRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAuthResponseSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "NotFound"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "logout"
                ],
                "summary": "Logout a user",
                "description": "Returns the user data",
                "operationId": "logout",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLogoutRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserLogoutResponseSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "NotFound"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "refresh"
                ],
                "summary": "Refresh token",
                "description": "Returns the user data",
                "operationId": "2f35a9fa078ee4c8122aad8b8c0adc33",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCheckTokenRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCheckTokenResponseSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCheckTokenResponseSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/authorize": {
            "post": {
                "tags": [
                    "authorize"
                ],
                "summary": "Check if user is logged in",
                "description": "Returns the user data",
                "operationId": "87962c23d9bbaf3f3b9d5e3eb545adb0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserAuthorizeRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAuthorizeResponseSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAuthorizeResponseSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/change-password": {
            "post": {
                "tags": [
                    "authorize"
                ],
                "summary": "Check if user is logged in",
                "description": "Returns the user data",
                "operationId": "e9210ecd43221a32527c555522eb5ade",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPasswordChangeRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserPasswordChangeResponseSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserPasswordChangeResponseSchema"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserAuthRequestSchema": {
                "title": "User authentication request model",
                "description": "User authentication request model",
                "properties": {
                    "username": {
                        "title": "Username",
                        "description": "Username of the user",
                        "type": "string",
                        "example": "tugrulyildirim"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password for authentication",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserAuthRequestSchema"
                }
            },
            "UserAuthResponseSchema": {
                "title": "User authentication response model",
                "description": "User authentication response model",
                "properties": {
                    "success": {
                        "title": "Success",
                        "description": "Success",
                        "type": "boolean",
                        "format": "boolean",
                        "example": "true"
                    },
                    "id": {
                        "title": "id",
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "username": {
                        "title": "Username",
                        "description": "Username for authentication",
                        "type": "string",
                        "format": "string",
                        "example": "tugrulyildirim"
                    },
                    "name": {
                        "title": "Name",
                        "description": "First name of the user",
                        "type": "string",
                        "format": "string",
                        "example": "Tuğrul"
                    },
                    "surname": {
                        "title": "Surname",
                        "description": "Surname of the user",
                        "type": "string",
                        "format": "string",
                        "example": "Yıldırım"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the user",
                        "type": "string",
                        "format": "string",
                        "example": "iletisim[at]tugrulyildirim.com",
                        "nullable": true
                    },
                    "access_token": {
                        "title": "Access Token",
                        "description": "Access Token",
                        "type": "string",
                        "format": "string",
                        "example": "token12313123..."
                    },
                    "token_type": {
                        "title": "Token Type",
                        "description": "Token Type",
                        "type": "string",
                        "format": "string",
                        "example": "Bearer"
                    },
                    "expires_at": {
                        "title": "Expires at",
                        "description": "Expires at",
                        "type": "integer",
                        "format": "int64",
                        "example": 3600
                    },
                    "is_active": {
                        "title": "Is it active?",
                        "description": "Is it active?",
                        "type": "boolean",
                        "format": "boolean",
                        "example": true
                    },
                    "privileges": {
                        "title": "Privileges",
                        "description": "All privileges of the user",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "type": "object",
                            "example": {
                                "name": "admin",
                                "value": true
                            }
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserAuthResponseSchema"
                }
            },
            "UserAuthorizeRequestSchema": {
                "title": "User authorisation request model",
                "description": "User authorisation request model",
                "properties": {
                    "token": {
                        "title": "Token",
                        "description": "Send token in header",
                        "type": "string",
                        "format": "string",
                        "example": "bearer token123..."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserAuthorizeRequestSchema"
                }
            },
            "UserCheckTokenRequestSchema": {
                "title": "User check token request model",
                "description": "User check token request model",
                "properties": {
                    "token": {
                        "title": "Token",
                        "description": "Token for check",
                        "type": "string",
                        "format": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiaXNfYWN0aXZlIjoxLCJ1c2VyX3R5cGUiOjIsImVtYWlsIjoidHVncnVsQGdtYWlsLmNvbSIsInVzZXJuYW"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserCheckTokenRequestSchema"
                }
            },
            "UserLogoutRequestSchema": {
                "title": "User logout request model",
                "description": "User logout request model",
                "properties": {
                    "token": {
                        "title": "Token",
                        "description": "Token for logout",
                        "type": "string",
                        "format": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiaXNfYWN0aXZlIjoxLCJ1c2VyX3R5cGUiOjIsImVtYWlsIjoidHVncnVsQGdtYWlsLmNvbSIsInVzZXJuYW"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserLogoutRequestSchema"
                }
            },
            "UserPasswordChangeRequestSchema": {
                "title": "User password change request model",
                "description": "User password change request model",
                "properties": {
                    "old_password": {
                        "title": "Old password",
                        "description": "Old password",
                        "type": "string",
                        "example": "123456"
                    },
                    "new_password": {
                        "title": "New password",
                        "description": "New password",
                        "type": "string",
                        "example": "123456"
                    },
                    "confirm_password": {
                        "title": "Confirm password",
                        "description": "Confirm new password",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserPasswordChangeRequestSchema"
                }
            },
...
            "UserAuthorizeResponseSchema": {
                "title": "User authorisation response model",
                "description": "User authorisation response model",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserAuthorizeResponseSchema"
                }
            },
            "UserCheckTokenResponseSchema": {
                "title": "User check token response model",
                "description": "User check token response model",
                "properties": {
                    "success": {
                        "title": "Success",
                        "description": "Success",
                        "type": "boolean",
                        "format": "boolean",
                        "example": "true"
                    },
                    "token": {
                        "title": "token",
                        "description": "User token",
                        "type": "string",
                        "format": "string",
                        "example": "bearer token123..."
                    },
                    "message": {
                        "title": "message",
                        "description": "Response message",
                        "type": "string",
                        "format": "string",
                        "example": "token_is_valid"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserCheckTokenResponseSchema"
                }
            },
            "UserLogoutResponseSchema": {
                "title": "User logout response model",
                "description": "User logout response mode",
                "properties": {
                    "success": {
                        "title": "Message",
                        "description": "Message of the response",
                        "type": "string",
                        "format": "boolean",
                        "example": "logout_success"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserLogoutResponseSchema"
                }
            },
            "UserPasswordChangeResponseSchema": {
                "title": "User password change response model",
                "description": "User password change response model",
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "Message of the response",
                        "type": "string",
                        "example": "password_change_success"
                    },
                    "success": {
                        "title": "Success",
                        "description": "Success of the response",
                        "type": "boolean",
                        "format": "boolean",
                        "example": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserPasswordChangeResponseSchema"
                }
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://admin.local/api/v1/auth/authorize",
                        "tokenUrl": "http://admin.local/api/v1/auth/token",
                        "refreshUrl": "http://admin.local/api/v1/auth/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}